Bug Fix: Fix bogus check for invalid clients
The CLIENT_VALID macro, which checks for valid clients, did not
check the fd of the socket, which resulted in the code using
invalid clients in portions of code, such as ws_ping()/
send_ping_close().

Fix this by checking if the connection fd is '-1', which means
a free slot.

Signed-off-by: Michał Gapiński <mike@gapinski.eu>